Fix old integration - Adapt with new refactor

+++ DataStructureVisualizer/include/Model/HeapStructure/IHeapStructure.h
@@ -3,20 +3,21 @@
 
 #include "Model/IVisualizable.h"
 #include <vector>
+#include <cstddef>
 
 class IHeapStructure : public IVisualizable {
     protected:
     std::vector<int> heapArray;
 
-    int parent(int n, int idx);
-    int left(int n, int idx);
-    int right(int n, int idx);
+    size_t parent(size_t n, size_t idx);
+    size_t left(size_t n, size_t idx);
+    size_t right(size_t n, size_t idx);
 
-    void shiftUp(int i, Timeline& timeline);
-    void fixNode(int i, Timeline& timeline);
-    void heapify(int n, int idx, Timeline& timeline);
+    void shiftUp(size_t i, Timeline& timeline);
+    void fixNode(size_t i, Timeline& timeline);
+    void heapify(size_t n, size_t idx, Timeline& timeline);
 
-    virtual bool compare(int idx1, int idx2) = 0;
+    virtual bool compare(size_t idx1, size_t idx2) = 0;
 
     public:
     void initialize(const std::vector<int>& data, Timeline& timeline);
